fix: add all available model fields to ensure that multiple models don't trigger additional fields check for csv
restructured the reader tests as well
